forbidden delete msg

+++ src/channel.js
@@ -1044,11 +1044,13 @@ Max.Channel.prototype.deleteMessage = function(messageID) {
     setTimeout(function() {
         if (!mCurrentUser) return def.reject(Max.Error.SESSION_EXPIRED);
         if (!self.name) return def.reject(Max.Error.INVALID_CHANNEL);
+        if (!self.isOwner()) return def.reject(Max.Error.FORBIDDEN);
         if (!messageID) return def.reject(Max.Error.INVALID_MESSAGE_ID);
 
         Max.Request({
             method: 'DELETE',
-            url: '/com.magnet.server/channel/message/' + messageID
+            url: '/com.magnet.server/channel/message/' + messageID,
+            isLogin: true
         }, function(res) {
             def.resolve(res.message, res.code);
         }, function() {
